only hide word suggestions when inline completions are enabled
