+++ src/vs/workbench/contrib/chat/browser/chatStatus/chatStatus.ts
@@ -4,24 +4,8 @@
  *--------------------------------------------------------------------------------------------*/
 
 import { ChatEntitlement, IChatEntitlementService } from '../../../../services/chat/common/chatEntitlementService.js';
-import { IConfigurationService } from '../../../../../platform/configuration/common/configuration.js';
-import product from '../../../../../platform/product/common/product.js';
-import { isObject } from '../../../../../base/common/types.js';
 
 export function isNewUser(chatEntitlementService: IChatEntitlementService): boolean {
 	return !chatEntitlementService.sentiment.installed ||					// chat not installed
 		chatEntitlementService.entitlement === ChatEntitlement.Available;	// not yet signed up to chat
-}
-
-export function isCompletionsEnabled(configurationService: IConfigurationService, modeId: string = '*'): boolean {
-	const result = configurationService.getValue<Record<string, boolean>>(product.defaultChatAgent.completionsEnablementSetting);
-	if (!isObject(result)) {
-		return false;
-	}
-
-	if (typeof result[modeId] !== 'undefined') {
-		return Boolean(result[modeId]); // go with setting if explicitly defined
-	}
-
-	return Boolean(result['*']); // fallback to global setting otherwise
 }
